I added an analytical solution to the inverse kinematics task.
Previously I used the gradient descent solution. But that one was slow
and inaccurate. Before the deadline I really wanted to also add an
analytical solution based on the B-Human Team Documentation. It works
better than the gradient approach but I have kept both since the
gradient approach took a lot of work.